refactor(folder): folder domain structure ajusted
