+++ resources/views/admin/ticket/create.blade.php
@@ -1,0 +1,3 @@
+<div>
+    <!-- I begin to speak only when I am certain what I will say is not better left unsaid. - Cato the Younger -->
+</div>
